Dialog Box is recreated on configuration change
